Add a unique nonce to yum repoids
This seems the simplest immediate solution to the problem that we need
to define two repos for the same content set, and the repoids are named
after the content set but need to be unique. So we track how many repos
are seen and append the number to each repoid.

Note that this does not strictly guarantee uniqueness; if someone
cleverly named two content sets with the right numbers at the end a
collision is *possible* but this will never happen in real life.