Experimental mirror input and ImageStream generator
Generates input file for 'oc image mirror' command. Used to
synchronize OCP content from internal registries to quay.

Generates a kubernetes ImageStream object to send to a release
controller which is responsible for creating update and install
payloads.

This is still experimental and VERY specific to ART
requirements. Requires future refactoring to be generally usable.