Fix sync input generating logic
Logic for adding items to the "src=dest" list was incorrectly outside
of the conditional that actually checked those conditions.

This line of code is now inside the conditional block. Now, only items
which exist in the source registry AND meet our naming conventions are
added to the 'oc image mirror' src=dest list.