distgit/runtime: improve Dockerfile blame feature
When we limited source checkouts to --depth 1 that caused the logic that
looked for the last committer of the Dockerfile to always be that last
commit. Typically this is a merge-bot anyway, and so we would see in
doozer logs:

Unable to get author email for last Dockerfile commit:
Last committer email found, but is not @redhat.com address:
openshift-merge-robot@users.noreply.github.com

With this change, we get the last 50 commits of the source, giving us a
decent chance of finding the real last commit if it hasn't been buried.
We also omit merges and initial add of the file (with a shallow
checkout, the "first" commit is credited with adding everything).