distgit,runtime: on scan-sources, check if a build exists
The existing logic only checked that sources had been successfully
rebased/released into the distgit. If not built successfully it would
not be included in incremental builds until the source changed again.

Updated logic looks for a build in brew matching the distgit contents.
If there isn't one, it checks the release timestamp to determine if it's
stale enough to be counted as changed for an incremental build. We don't
want to spam the image owners about repeated failures but we also don't
want to ignore a lost build too long; this threshold is configurable.

This also handles no-source distgits, though imperfectly. If the
Dockerfile gets an updated release which is not built, and it is past
the threshold, it also is considered changed for an incremental build.
However, if someone manually updated the distgit content without
changing the dockerfile release label, it would not be considered
changed.