distgit.py: search builds for component not dg name
When doing a scan for changed sources, the test for whether the distgit
copy of the source has been built looks for brew builds. But it was
looking for the distgit name instead of the component. For RPMs they are
the same but images generally add -container or -apb or explicitly
specify a component.