Run `tox` on Travis
My initial implementation of running tests on Travis was rather opinionated,
assuming that people would only use docker/podman to run the tests.

On this PR I'm making Travis run `tox` inside a container, but during development
we have a choice of running either inside the container or directly on the host.

    $ podman build -f .ci/Dockerfile -t doozer-ci-env .
    $ podman run -it --rm -v $(pwd):/code -w /code doozer-ci-env

or simply:

    $ tox

Related ticket:
https://jira.coreos.com/browse/ART-547