Python 3 porting: Run `futurize --stage1`
I am trying [Futurize][] for porting Python 2 code to Python 3.
The tool is one of recommended tools listed on the official Python HOWTOs documentation [Porting Python 2 Code to Python 3][].

Futurize can divide the porting into two stages. Stage 1 is for “safe” changes that modernize the code but do not break Python 2.6 compatibility or introduce a depdendency on the future package. Stage 2 is to complete the process.

The PR is generated with the following command:

```bash
futurize --stage1 --unicode-literals --nobackups --write .
```

[Porting Python 2 Code to Python 3]: https://docs.python.org/3/howto/pyporting.html
[Futurize]: http://python-future.org/automatic_conversion.html

Only stage 1 of the Futurize porting process is run so that we are expecting this will not break our existing Python 2 compatibility but it will still not run on Python 3.