release_calc_previous: Set minimum versions based on build-suggestions
We now have a mandate to ensure that z-stream upgrades are cumulative.
Overall it is also better for us to always offer and upgrade path that's
directly to the best version available. So rather than relying on a rolling
window of ten most recent z-streams and the two most recent previous minor
versions we'll use a static minimum versions and a block list sourced from OTA
managed content.

Prior to this change

$ ./doozer release:calc-previous --version 4.5.8
4.5.8,4.5.7,4.5.6,4.5.5,4.5.4,4.5.3,4.5.2,4.5.1,4.5.0-0.hotfix-2020-08-24-185832,4.4.20,4.4.19,4.4.18

After this change
$ ./doozer release:calc-previous --version 4.5.8
4.5.8,4.5.7,4.5.6,4.5.5,4.5.4,4.5.3,4.5.0-0.hotfix-2020-08-24-185832,4.4.20,4.4.19,4.4.18,4.4.17,4.4.16,4.4.15,4.4.14,4.4.13

Based on the following metadata
$ curl https://raw.githubusercontent.com/sdodson/cincinnati-graph-data/suggestions/build-suggestions/4.5.yaml
```
---
default:
  minor_floor: 4.4.13
  minor_block_list: []
  z_floor: 4.5.3
  z_block_list: []
s390x:
  minor_floor: 4.4.24
  minor_block_list: []
  z_floor: 4.5.3
  z_block_list: []
```