ART-1080: update doozer:scan-sources to note config changes
For each image metadata, generate a digest that captures all relevant config. The current relevant items for images are:

- the metadata config itself (remove owners: though)
- the entries in group.yml attr repos: corresponding to the image's enabled_repos and non_shipping_repos attrs
- the entries in streams.yml corresponding to any stream: entries in the image from and from.builder attrs

When rebasing, create a dict with these entries, serialize it as json (with ordered keys so it's consistent), and store a digest of that in new distgit file .oit/config_digest.

When scan-sources runs, generate the digest the same way and compare with the distgit. If the distgit digest is different or missing, the image should be rebuilt.