[ART-2816]: New implementation for building rpms
Introduces a new implementation for building rpms.

Per https://issues.redhat.com/browse/ART-2690, we want to get rid of
`tito` in Doozer rpm build to gain more control over the build process.

This PR:

- provides a new command `beta:rpms:build` for building rpms without
  tito, which can eventually replace `rpms:build` once tested;
- includes upstream git commit hash in RPM NVR. e.g. `openshift-4.8.0-202011111040.p0.git.abcdef1.el8`;
- uses asyncio as much as possible to run parallel tasks;
- refactors some code to reuse some existing logic;
- adds more unit tests.

This PR makes a few changes to the existing code, but it shouldn't break
the existing implementation.