ART-2823: update doozer release:gen-payload to sync split images
Be aware of the new `payload_name` field in image meta.
When syncing images, arch-specific image with explicit payload_name
serves as an override (https://issues.redhat.com/browse/ART-2823).

This PR also fixes a type error when `images_missing_builds` is not
empty, and changes the stand-in image from `cli` to `pod`.