[ART-3075] Add plashet subverb "for-assembly"
This adds plashet subverb "for-assembly" to creates a plashet containing arch specific yum repository subdirectories based on a complete set of RPMs required for image builds.
When doozer runs brew builds for an assembly image, it should only need to pass in a single repository created by for-assembly in order for the image to build successfully.

Implemented use cases:
1. The subverb is run against a runtime.assembly that does have a basis event.
2. The subverb is run against an assembly with a basis event and an "--image" argument.

Running against an assembly with a basis event and an "--rhcos" argument will be covered in a separate PR.