ART-3050: full assembly awareness for plashet from-tags (#439)
* ART-3050: full assembly awareness for plashet from-tags

When an artist is trying to refine an assembly with a basis event, existing assembly-specific builds for that assembly are not canonically part of the assembly until the artist pins then in the assembly metadata with "is". It follows that plashet should honor "is" when set for RPMs in the assembly metadata when constructing a repo.

Another change in behavior is that plashet should honor group "dependencies" defined in the assembly configuration. 431 takes care of determining what a group configuration should look like in light of a given assembly name on the command line. Unlike normal group.yml, assembly group definitions in releases.yml can add "dependencies" to the loaded models.

* Apply suggestions from code review

Co-authored-by: Justin Pierce <jupierce@redhat.com>

* config:plashet from-tags: Extracts some logic to a separate module

So that future development can reuse it.

Also limit the NVRs for signing to builds from tags and those pinned by
"is", because we are not able to sign non-OCP RPMs, which should already
be signed.

Co-authored-by: Justin Pierce <jupierce@redhat.com>