Have --force-yum-updates update only the final stage
The original intent here was to allow an artist to pin an assembly group
or assembly image dependency in releases.yml and have builder stages
in the Dockerfile install those updated packages.

In practice, builder images are diverse and https://coreos.slack.com/archives/GDBRP5YJH/p1626358901421900?thread_ts=1626187378.370800&cid=GDBRP5YJH
shows at least one instance where the builder image could not be updated
with "yum -y update" even with access to all of rhel8. The reason being
that a -devel package installed in the builder and that package was not
available in RHEL.

So the plan is, for now, to eliminate the ability to update a builder
image using assembly dependency pinning. We could accomplish this in the future
by only enabling the rebuild override repo for builder stages (i.e.
exposing ONLY the pinned dependencies indicated by the artist (instead
of all of RHEL). Since this could be tricky to accomplish elegantly (e.g
it would seem to require disabling non-override repos in build stages
prior to running yum -y update), we will instead move ahead without
this feature.

If an artist truly needs to override a specific package in a builder
image for an assembly, they can rebuild the builder images with the
desired package and then pin that builder image as an override in the
assembly definition.