Prepare for release v2.0.2
e810dcb Permit assemblies to use custom rhcos builds The RHCOS pipeline, when configured for a custom build, will bucket the resulting metadata files in s3 in a directory suffixed with `-custom`. In order to for assemblies to use custom RHCOS builds, we must try to find metadata in either the standard location or `-custom` location.
0a4d5ca [ART-3109] Use container.yaml to apply floating tags (#524)
e3e6315 ART-3437: Make bundle tag names configurable
36b3291 Prevent leading zero in semver (#527)
57af5c8 Revert "make bundle tag names configurable (#532)" (#533)
3a14f91 make bundle tag names configurable (#532)
3c16d82 Allow olm-bundle:print to skip when builds are missing
b3cc4c3 We are in the future
69e613e Update code to work with flexmock 0.11
4107d82 ART-2519 add check upstreampulls function (#488)
56325c6 bump-doozer: add annotated tag