get-nightlies: revamp for deeper comparison
Comparing timestamps on nightlies can only go so far.
For a more reliable indicator that all the nightlies selected are "the
same thing" get-nightlies now compares contents quite deeply.
This should be reliable enough to feed into gen-assembly or to use in
jobs checking that nightlies are available for upcoming releases.

It also allows specifying a nightly you'd like in the set and other
improvements.