Imagestream fixes (#665)
Only open one reconcilation PR at a time if component is using fallback branch
Prior to this change, a component that only has a 'master' branch, for example,
would have that branch used as a fallback for all active releases. This
made imagestreams open a PR against the master branch for reconciliation.
In these cases, we should only open a reconciliation PR against master when
the current group matches what release CI is tracking for master branches.