Support patch_version defined in assembly config
If we have assembly config like this:

```yaml
  art5142:
    assembly:
      type: custom
      basis:
        assembly: 4.12.99
        patch_version: 99
```

The release name will become `4.12.99-assembly.art5142`.

In case patch_version is not defined, let our automation to look at the chain of assembly inheritance to determine one, e.g.

```yaml
  art0000:
    assembly:
      type: custom
      basis:
        assembly: 4.12.88
  art0001:
    assembly:
      type: custom
      basis:
        assembly: art0000
  art0002:
    assembly:
      type: custom
      basis:
        assembly: art0001
```

The patch_version of art0002 will default to 88 because release 4.12.88 is the closest ancestor of type standard.