rhcos buildfinder: Find valid releases
Currently for 4.14, the rhcos jobs for the arches have resulted in good
builds, yet, the `release` job failed. This leads to the circumstance
that the builds are listed in `builds.json`, but are not viable for use.
In particular, the `oscontainer` key is missing from the metadata.

This commit ensures that the newest rhcos is selected where also this
label is set.